+++ R/find_formula.R
@@ -45,6 +45,9 @@
 #'  - `correlation`, for models with correlation-component like
 #'    `nlme::gls()`, the formula that describes the correlation structure
 #'
+#'  - `scale`, for distributional models such as `mgcv::gaulss()` family fitted
+#'    with `mgcv::gam()`, the formula that describes the scale parameter
+#'
 #'  - `slopes`, for fixed-effects individual-slope models like
 #'    `feisr::feis()`, the formula for the slope parameters
 #'
@@ -159,15 +162,20 @@ find_formula.gam <- function(x, verbose = TRUE, ...) {
   if (!is.null(f)) {
     if (is.list(f)) {
       mi <- .gam_family(x)
-      if (!is.null(mi) && mi$family == "ziplss") {
-        # handle formula for zero-inflated models
-        f <- list(conditional = f[[1]], zero_inflated = f[[2]])
-      } else if (mi$family == "Multivariate normal") {
-        # handle formula for multivariate models
-        r <- lapply(f, function(.i) deparse(.i[[2]]))
-        f <- lapply(f, function(.i) list(conditional = .i))
-        names(f) <- r
-        attr(f, "is_mv") <- "1"
+      if (!is.null(mi)) {
+        f <- switch(mi$family,
+          ziplss = list(conditional = f[[1]], zero_inflated = f[[2]]),
+          # handle formula for location-scale models
+          gaulss = list(conditional = f[[1]], scale = f[[2]]),
+          # handle formula for multivariate models
+          `Multivariate normal` = {
+            r <- lapply(f, function(.i) deparse(.i[[2]]))
+            f <- lapply(f, function(.i) list(conditional = .i))
+            names(f) <- r
+            attr(f, "is_mv") <- "1"
+            f
+          }
+        )
       }
     } else {
       f <- list(conditional = f)
@@ -517,10 +525,10 @@ find_formula.afex_aov <- function(x, verbose = TRUE, ...) {
     dv <- attr(x, "dv")
     id <- attr(x, "id")
 
-    within <- names(attr(x, "within"))
-    within <- paste0(within, collapse = "*")
-    within <- paste0("(", within, ")")
-    e <- paste0("Error(", id, "/", within, ")")
+    within_variables <- names(attr(x, "within"))
+    within_variables <- paste0(within_variables, collapse = "*")
+    within_variables <- paste0("(", within_variables, ")")
+    e <- paste0("Error(", id, "/", within_variables, ")")
 
     between <- names(attr(x, "between"))
     if (length(between) > 0L) {
@@ -528,10 +536,10 @@ find_formula.afex_aov <- function(x, verbose = TRUE, ...) {
       between <- as.character(tempf)[3]
       between <- paste0("(", between, ")")
 
-      within <- paste0(c(within, between), collapse = "*")
+      within_variables <- paste0(c(within_variables, between), collapse = "*")
     }
 
-    out <- list(conditional = stats::formula(paste0(dv, "~", within, "+", e)))
+    out <- list(conditional = stats::formula(paste0(dv, "~", within_variables, "+", e)))
     class(out) <- c("insight_formula", "list")
     out
   }
@@ -586,12 +594,10 @@ find_formula.gls <- function(x, verbose = TRUE, ...) {
   fcorr <- x$call$correlation
   if (is.null(fcorr)) {
     f_corr <- NULL
+  } else if (inherits(fcorr, "name")) {
+    f_corr <- attributes(eval(fcorr))$formula
   } else {
-    if (inherits(fcorr, "name")) {
-      f_corr <- attributes(eval(fcorr))$formula
-    } else {
-      f_corr <- parse(text = safe_deparse(fcorr))[[1]]
-    }
+    f_corr <- parse(text = safe_deparse(fcorr))[[1]]
   }
   if (is.symbol(f_corr)) {
     f_corr <- paste("~", safe_deparse(f_corr))
@@ -600,12 +606,10 @@ find_formula.gls <- function(x, verbose = TRUE, ...) {
   }
 
   l <- tryCatch(
-    {
-      list(
-        conditional = stats::formula(x),
-        correlation = stats::as.formula(f_corr)
-      )
-    },
+    list(
+      conditional = stats::formula(x),
+      correlation = stats::as.formula(f_corr)
+    ),
     error = function(x) {
       NULL
     }
@@ -1267,6 +1271,16 @@ find_formula.sem <- function(x, verbose = TRUE, ...) {
 #' @export
 find_formula.lme <- function(x, verbose = TRUE, ...) {
   fm <- stats::formula(x$terms)
+  .find_formula_nlme(x, fm, verbose = verbose, ...)
+}
+
+#' @export
+find_formula.glmmPQL <- function(x, verbose = TRUE, ...) {
+  fm <- stats::formula(x)
+  .find_formula_nlme(x, fm, verbose = verbose, ...)
+}
+
+.find_formula_nlme <- function(x, fm, verbose = TRUE, ...) {
   fmr <- eval(x$call$random)
   if (!is.null(fmr) && safe_deparse(fmr)[1] == "~1") {
     check_if_installed("nlme")
@@ -1276,12 +1290,10 @@ find_formula.lme <- function(x, verbose = TRUE, ...) {
   fcorr <- x$call$correlation
   if (is.null(fcorr)) {
     fc <- NULL
+  } else if (inherits(fcorr, "name")) {
+    fc <- attributes(eval(fcorr))$formula
   } else {
-    if (inherits(fcorr, "name")) {
-      fc <- attributes(eval(fcorr))$formula
-    } else {
-      fc <- parse(text = safe_deparse(fcorr))[[1]]$form
-    }
+    fc <- parse(text = safe_deparse(fcorr))[[1]]$form
   }
 
   f <- compact_list(list(
@@ -1364,12 +1376,10 @@ find_formula.BBmm <- function(x, verbose = TRUE, ...) {
 #' @export
 find_formula.mmclogit <- function(x, verbose = TRUE, ...) {
   f <- tryCatch(
-    {
-      list(
-        conditional = stats::formula(x),
-        random = stats::as.formula(parse(text = safe_deparse(x$call))[[1]]$random)
-      )
-    },
+    list(
+      conditional = stats::formula(x),
+      random = stats::as.formula(parse(text = safe_deparse(x$call))[[1]]$random)
+    ),
     error = function(x) {
       NULL
     }
@@ -1418,12 +1428,10 @@ find_formula.stanreg <- function(x, verbose = TRUE, ...) {
     # special handling for stan_gamm4
     if (inherits(x, "gamm4")) {
       f.random <- tryCatch(
-        {
-          lapply(.findbars(stats::formula(x$glmod)), function(.x) {
-            f <- safe_deparse(.x)
-            stats::as.formula(paste0("~", f))
-          })
-        },
+        lapply(.findbars(stats::formula(x$glmod)), function(.x) {
+          f <- safe_deparse(.x)
+          stats::as.formula(paste0("~", f))
+        }),
         error = function(e) {
           NULL
         }
@@ -1484,8 +1492,8 @@ find_formula.MCMCglmm <- function(x, verbose = TRUE, ...) {
 find_formula.BFBayesFactor <- function(x, verbose = TRUE, ...) {
   if (.classify_BFBayesFactor(x) == "linear") {
     fcond <- utils::tail(x@numerator, 1)[[1]]@identifier$formula
-    dt <- utils::tail(x@numerator, 1)[[1]]@dataTypes
-    frand <- names(dt)[which(dt == "random")]
+    dat_types <- utils::tail(x@numerator, 1)[[1]]@dataTypes
+    frand <- names(dat_types)[which(dat_types == "random")]
 
     if (is_empty_object(frand)) {
       f.random <- NULL
@@ -1816,15 +1824,14 @@ find_formula.model_fit <- function(x, verbose = TRUE, ...) {
     fc <- try(.formula_clean(f[[1]]), silent = TRUE)
     if (inherits(fc, "try-error")) {
       format_error(attributes(fc)$condition$message)
-    } else {
-      if (verbose) {
-        format_warning(paste0(
-          "Using `$` in model formulas can produce unexpected results. Specify your model using the `data` argument instead.", # nolint
-          "\n  Try: ", fc$formula, ", data = ", fc$data
-        ))
-      }
-      return(FALSE)
     }
+    if (verbose) {
+      format_warning(paste0(
+        "Using `$` in model formulas can produce unexpected results. Specify your model using the `data` argument instead.", # nolint
+        "\n  Try: ", fc$formula, ", data = ", fc$data
+      ))
+    }
+    return(FALSE)
   }
   return(TRUE)
 }